fix(mobile): keep in-app account deletion after the android split
